mixed up .h + .c
