+++ main/CMakeLists.txt
@@ -2,8 +2,8 @@ idf_component_register(
   SRCS
   "isotp_ble_bridge_main.c"
   "isotp.c"
-  "wifi_server.c"
   "ble_server.c"
+  "wifi_server.c"
   "web_server.c"
   "ws2812_control.c"
   INCLUDE_DIRS "."
