+++ main/wifi_server.c
@@ -2,10 +2,9 @@
 #include "esp_log.h"
 #include "esp_event.h"
 #include "esp_wifi.h"
-
 #include "wifi_server.h"
 
-static const char *TAG = "wifi softAP";
+static const char *TAG = "wifi_server";
 
 static void wifi_event_handler(void* arg, esp_event_base_t event_base, int32_t event_id, void* event_data) {
     if (event_id == WIFI_EVENT_AP_STACONNECTED) {
